[FIX] documents_project: group by stage_id in my tasks in portal
Before this commit, when the user has more than one project linked to a
same stage then a traceback will be raised when the user will group by
stage in `/my/tasks` portal view.

This commit reviews the visibility condition of the Documents button in
that portal list view and makes sure the project variable in the
template contains either no project or just one project to avoid having
a traceback because `project` variable contains more than one project.
And so, in `/my/tasks` view, the Documents button should not be
displayed when the view is grouped by Stage.

Steps to reproduce the issue
============================

1. Install documents_project
2. Create 2 projects
3. Go to Tasks stage menu and set the both projects in the first stage
   (create new stage if there is no stage displayed in that menu)
4. create a task for each project with that first stage
5. Go to `/my/tasks` portal list view
6. Group by `Stage`

Actual Behavior
---------------

A traceback is raised because there is more than one project contained
inside `project` variable.

Expected Behavior
-----------------

The documents button should not be displayed in `/my/tasks` since we
could have more than one project in that view.

task-4239772

X-original-commit: d4ee0e5c040f6840058e27c2939ff48415b7bf8e
Part-of: odoo/enterprise#72369
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>