[FIX] documents: drags to trash reload search panel
Purpose
=======
Fix the left panel which wasn't updated when dragging a folder to trash.

Specification
=============
When dragging a document of type 'folder' to the trash, the model is
correctly being notified but the search model isn't, the folder is thus
still visible in the left panel.
Fixing that by also reloading the search model when archiving a document.

Moving the reloading methods calls from toggleArchive to onDropDocuments
as it is the only method where we don't reload the model and search model
after archiving. Keeping the reload calls in toggleArchive caused redundant
reloads of the models.

Task-4260645

closes odoo/enterprise#73440

X-original-commit: 971e225c85508edc5baa0735e79b278fbf57703b
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Signed-off-by: Amélie Dieudonné (amdi) <amdi@odoo.com>