[FIX] account_reports: previous options traceback
Before this commit when adding value in the ec_tax_filter_selection dict in
an extends of the method, the previous options would be kept and this will lead
to a traceback because the key would not be recognised in the other report.

For example, in the l10n_si localisation we add twos keys into the dictionary,
and then we go back on the generic one, the two keys would still be in the
previous options and so it would be passed in the ec_tax_filter_selection and
crash.

closes odoo/enterprise#73796

Task: 3901247
X-original-commit: f0e1725b368c0ffd65240f8ece94b4155ce50eeb
Related: odoo/odoo#187201
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>