[IMP] l10n_ar_reports: VAT Book: use load_more_limit
When this report contained too many lines (~6000), the browser couldn't handle the load and did not manage to render it. This is a known case, which normally only happens with line expansion, for which the prefix groups and load more features have been made. In this case, though, it's a bit different, as those lines are directly at the root of the report (they don't come from a line expansion). We make the choice here to use the load_more_limit anyway as a means to put a limit to the number of lines that are shown.

When this limit is reached, we just don't display the exceeding lines, and replace them by a summary line.

To avoid confusion, a warning is now shown on top of the report when some lines need to be hidden. Since such a warning requires a module update (because it's a template, and needs to exist in db), we only apply our new trimming mechanism if the warning template exists (and can hence be displayed). This way, we won't change the behavior on existing databases without being sure the user is properly notified. Plus, thanks to that, we're sure the old load_more_limit value (originally 80 by default) set on the VAT Book will have been increased to a value fitting our use case more, so we won't bother people with excessive trimming.

Original fix suggestion raised here by Adhoc: https://github.com/odoo/enterprise/pull/45936

closes odoo/enterprise#73025

X-original-commit: e861da91abde1a26502f2be13d5b23b1137a6935
Signed-off-by: Julien Van Roy (juvr) <juvr@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>