[FIX] l10n_ae_report,l10n_ae_corporate_tax_report: module installation
Before this commit, when installing the account_reports and l10n_ae modules, the
l10n_ae_reports and corporate_tax_report modules were not installed, even
though  they depended on these modules. This caused an issue during an account
report test, as the necessary field was not created, leading to an "Invalid
field"  error: ('l10n_ae_tax_report_liabilities_account' on model 'res.company')

By removing the countries field from the manifest, the modules now auto-install
as expected, and the test no longer fails. But why? If a module depends on
another module that has the countries attribute, there’s no need to include it
back; the auto-installation will only be triggered if there is a company that
belongs to that country.

closes odoo/enterprise#72158

X-original-commit: 070e5f53e8a8376854158a5aff78b99e5e34b5dc
Signed-off-by: Quentin De Paoli (qdp) <qdp@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>