[FIX] account_accountant: edit exchange in bank reco widget
Let's say you have:
- One statement line of 3000 USD / 1000 MXN (rate 1:3)
- Three invoices of 1000 USD / 500MXN (rate 1:2)

=> If you reconcile in the bank widget, the statement entry amounts will be
divided by 3 ending with 3 lines of 1000 USD and 333.33 MXN leaving a 0.01
unmatched. This can be an issue for the validation of payment EDI for
localyzations like Mexican one.

This PR gives more flexibility to user when dealing when bank reconciliation in
multiple currencies by allowing to edit amounts on generated exchange moves to
allocate this small difference.

opw-4128899

closes odoo/enterprise#74372

X-original-commit: c76ae9249bd4c24ef0be6b8237a34c4fb08d4f46
Signed-off-by: Laurent Smet (las) <las@odoo.com>