[REF] point_of_sale: update links when setting object value
In this commit, we are now instantiating a record as a Proxy that traps the
setter. When setting a value for a certain key on a record, we call the `update`
internally to ensure that related records are linked. This means that developers
are no longer required to use `record.update` to apply changes on a record, they
can just set the value as if the record is a normal object.

closes odoo/enterprise#72198

Related: odoo/odoo#184108
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>