[FIX] iot: add missing sudo calls to `_default_token`
Before this change, if a regular user was granted access
to the IoT app, they would get a traceback on trying to
pair an IoT box as they cannot access `ir.config_parameter`.
This PR adds the missing `sudo` calls to fix this traceback
and allow the user to pair an IoT box.

task-4216519

closes odoo/enterprise#72498

X-original-commit: c9069409760f540a9ecfebeadcecb1e1d3ed393b
Signed-off-by: Quentin Lejeune (qle) <qle@odoo.com>