[FIX] account_reports: journal filter crash when no journal is available on the company
When we tried to open the report with a company without any journal and that the report had the journal filter set, it would crash.

The goal is to set the label of the filter explicitely to "No Journal" and make the filter unclickable when this is the case.

task-4267503

closes odoo/enterprise#72657

X-original-commit: 143488b03d44d481b48607eb2801876bea5ab41c
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Logan Staelens (lost) <lost@odoo.com>