[IMP] knowledge: remove editor command "ADD_EXTERNAL_STEP"
This command's purpose was to dispatch the event at each individual external
step, but it was only used for embedded components, and `onExternalHistorySteps`
resource (which handles every batch of collaborative steps) is enough for
embedded components, and it is called less often. This commit therefore removes
`ADD_EXTERNAL_STEP` command.

task-4281356

closes odoo/enterprise#72778

X-original-commit: 69df97593d36d9cd9af27fd0427e855a439178fa
Related: odoo/odoo#185274
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>