[IMP] timesheet_grid: add widget to getPropertyFieldInfo function
Until now, if a widget needed to be specified for the renderer, a hack
was needed, the widget needed to be set as a field type. The task
4224192 aim to add a warning when using a widget on a wrong field (field
type) and the current hack will raise the warning.

This commit adds widget, a new parameter, to the getPropertyFieldInfo
function. This parameter, will allow specifying the widget that we want
to use on the fieldInfo.

closes odoo/enterprise#72256

Task-id: 4224192
Related: odoo/odoo#184244
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>