[FIX] stock_barcode: recognise location scan when it is proposed
Steps to reproduce:

- Enable "Multi-Steps Routes" in the settings
- Add a barcode to a different location that WH/Stock
- Go to the barcode > Operations > Internal Transfers > New
> The header propose to scan a product
- Do not scan the product by click on `+ ADD PRODUCT`
- Add 1 unit of your favorite product
> The header propose to scan a product or a destination location
- Scan the location you had set
> The location scan is ignored

Cause of the issue:

During the `_parseBracode` called to determine what was scanned by the
barcode, a location corresponding to the barcode scanned will be found.
To use this barcode, if it makes sense, the `_setLocationFromBcode` will
be called in turn. However, during internal transfers, this location
will only be used as a destination location under very specific
conditions:
https://github.com/odoo/enterprise/blob/8035abddf9991f6bfb71bb7889b703e5ed31a179/stock_barcode/static/src/models/barcode_picking_model.js#L1603-L1619
And, since we did add the product manually rather than using a scan, in
our case, `this.previousScannedLines.lengt` will be null rather than
positive so that the location will not be used to set the destination.

opw-4201489

closes odoo/enterprise#73267

X-original-commit: b9b45e50cc3a2a12e62183035de96eaa1d529cc8
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>