[FIX] helpdesk_fsm: not create an empty partner when no info
Steps:
Create a ticket with no partner.
Click on 'Plan Intervention'.
The modal opens.

Issue:
See that a partner "@" has been created and set.

Cause:
Clicking on the button calls `action_generate_fsm_task`,
which calls `_find_or_create_partner`,
but the args, `self.partner_name` and `self.partner_email`, are empty.

Solution:
Not try to create a partner when the fields both empty.
Also, if there was no partner on the ticket,
it takes the value of the wizard's partner when saving the wizard
(clicking on "Create (& View) Task".

task-4265584

closes odoo/enterprise#73107

X-original-commit: f655a24a4891cff104c0b1e7540b57c43b902e07
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>