[FIX] stock_barcode: split uncomplete lines when exiting barcode
steps to reproduce:

- Create a delivery order for 5 units of a product
- Scan 1 unit and click on the pencil icon to edit the record further
- On the digipad access the backend (for instance by clicking the
  external link arrow next to the location).
- Come back to the record
> The line now displayed a 1/1 but does not records what remains: 0/4.

Cause of the Issue:

Setting a quantity of 1 on the move line will trigger a write call on
the related stock move line to update its quantity. However, when you
leave the barcode, even thought it will save these change, it will not
split the move in 2 for the barcode app to be able to rely on the
backend data to generate a correct sate of the record. Threfore, when
you come back to the record, the main component will be regenerated from
the backend values of the `get_barcode_data`:
https://github.com/odoo/enterprise/blob/7b28f4bbd3c20559e529c3134e5b65b089b78806/stock_barcode/static/src/components/main.js#L76-L80
And will onlydisplay the "1/1" line that was reported from the back end.

Fix:

The `split_uncompleted_moves` method taking care of splitting the moves
for barcode data compatibility:
https://github.com/odoo/enterprise/blob/7b28f4bbd3c20559e529c3134e5b65b089b78806/stock_barcode/models/stock_move.py#L9-L11
is already called when you exit the barcode via the exit button:
https://github.com/odoo/enterprise/blob/7b28f4bbd3c20559e529c3134e5b65b089b78806/stock_barcode/static/src/components/main.js#L228-L231
https://github.com/odoo/enterprise/blob/7b28f4bbd3c20559e529c3134e5b65b089b78806/stock_barcode/static/src/models/barcode_picking_model.js#L762-L765
These were added by commit 2eb6465e4f18c701efdf28ea4b8f95f3119d3db8
However, it should IMO be called each time you leave the app that is
just before destroying the component. Furthermore, note that it will
only split the moves if a save happend (which is the expacted behavior).

opw-4232106

closes odoo/enterprise#74166

X-original-commit: 6cc751fddc3c243b91187011c6d9c080ada910b7
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>