[FIX] documents: fix unscoped `z-index` property
This commit fixes an issue about a `z-index` CSS property defined
in Documents that is affecting the whole back-end and front-end
environment as it is not scoped within a Documents related selector.

While this CSS is actually needed if you have a dropdown with a `+ X`
item at the end to prevent the button from being rendered under the view,
it needs to be scoped to Documents only. We also take that opportunity to
use the `SCSS` variable related to that `z-index` value.

task-4270185

closes odoo/enterprise#73169

X-original-commit: 4f79287c2e413366f1bb05c5b15f6855b4195670
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Chrysanthe Gomrée (chgo) <chgo@odoo.com>