[FIX] l10n_mx_edi: conditionally display "Predial Account" column in invoice PDF
The "Predial Account" column appeared in all signed invoices, regardless of whether the invoice lines included rental products with a predial account. This resulted in a non-organized layout with an empty column for most invoices.

Steps to reproduce:
1. Install the l10n_edi_mx module.
2. Create and sign an invoice with any product and customer.

Expected behavior:
The "Predial Account" column should only be displayed when the invoice contains products with a predial account.

Actual behavior:
The "Predial Account" column is displayed in all invoices, even when it's empty.

opw-4233358

closes odoo/enterprise#74125

X-original-commit: 0d0793064215bfe75cc070b04b4cb240b91769d9
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>
Signed-off-by: Abdelaziz EL HAYYANY (azey) <azey@odoo.com>