[IMP] appointment: remove widget date on datetime field
This commit removes the date widget on datetime fields. The widget is
ignored if used on a datetime fields, see the tests : `DatetimeField
with date/datetime widget (with day change) does not care about widget`
and `DatetimeField with date/datetime widget (without day change) does
not care about widget.`

This commit is part of a bigger task, that aim to raise an error if a
widget is used in an unsupported field type, in this case the widget
date, will only support the date type fields.

closes odoo/enterprise#72374

Task-id: 4224192
Signed-off-by: Michaël Mattiello (mcm) <mcm@odoo.com>