[PERF] documents_project: don't count docs in project's kanban view
Description:
Following a32825ee00f2b330d99113f4d8c1488903fe744e, the "Documents"
menu-item is hidden when there are no documents in it. The presence
of the field `document_count` in the view triggers
`_compute_documents`, which is relatively slow for projects with a
lot of tasks.
It's not worth to incure the cost at each load of this view (which
is a hot path) just for hiding a button in a sub-menu.

Spec Change:
The menu-item "Documents" is present even if there are no document
and you have differnet than `'none'` permission on the folder.

Benchmark:
Loading "Projects" kanban view with 5 favourite projects, one of
which has 1M+ tasks on next.odoo.com

|               | Before | After | Speedup |
|---------------|--------|-------|---------|
| Timings (hot) | 10.2s  | 790ms | 12.9x   |

closes odoo/enterprise#73103

X-original-commit: f004d21923e375486750f55c68ebf2837eb4c8a0
Signed-off-by: Pierre Masereel (pim) <pim@odoo.com>
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>