[IMP] account_reports,*: remove 'disabled' company filter
*l10n_es_reports,l10n_uk_reports_cis

The 'disabled' value of the multi-company filter is used to tell the
report it should only consider the active company (and its branches),
ignoring all other companies that would be selected. Though, it's not
widely used, and the reports using it probably might run just as well
with 'selector'.

Also, it can create a bit of confusion, since it's not obvious to the
user which report uses or doesn't use the selector.

In the related enterprise commit, we simplify this by removing the
'disabled' value for the multi-company filter.

This commit adapt the existing code to this change.

task-4260664

closes odoo/enterprise#72402

Related: odoo/odoo#184553
Related: odoo/upgrade#6646
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>